requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.38.22
botocore==1.38.22
# celery==5.5.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2.1
# django-celery==3.1.17
django-cors-headers==4.7.0
django-environ==0.12.0
django-jalali==7.4.0
django-parler==2.3
django-parler-rest==2.2
# django-redis==5.4.0
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
humanize==4.12.3
jalali_core==1.0.0
jdatetime==5.2.0
jmespath==1.0.1
# kombu==5.5.3
pillow==11.2.1
prompt_toolkit==3.0.51
PyJWT==2.9.0
python-dateutil==2.9.0.post0
rosetta==0.3
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
# vine==5.1.0
wcwidth==0.2.13
# celery>=5.3.0
# redis>=5.0.0
# django-celery-beat>=2.5.0
django-filter>=23.1
psycopg2>=2.9.10
pytz>=2025.2
gunicorn>=23.0.0